Updated user level display.

+++ phase2/app/src/main/java/com/example/phase1activity/ui/matching_game/MatchingGamePresenterImpl.java
@@ -109,7 +109,7 @@ public class MatchingGamePresenterImpl implements MatchingGamePresenter {
       }
 
       if (matchesToBeMadeAfter == 0) {
-        double score = manager.getScore();
+        int score = manager.getScore();
         String statDisplayText = SCORE + score;
         view.setDisplayStat(statDisplayText);
         view.updateNextLevelButton();
